refactor KubernetesResourceHandler to explicitly take components not the whole charm
Previous version accepted a `CharmBase` object that conformed to an interface (CharmBase with/self.template_files, etc).  This left the KubernetesResourceHandler and Charm class tightly coupled.  This change refactors to explicitly accept factories for the inputs needed from the Charm class (eg: a function that creates template_files when needed) rather than use attributes directly on a Charm class